# roast-fe/api_doc.yaml
openapi: 3.0.3
info:
  title: Roast Anything API
  version: 1.0.0
  description: Backend API for the Roast Anything AI roasting platform.
paths:
...
  /api/v1/roasts/{id}/:
    get:
      operationId: roasts_retrieve
      description: |-
        GET    /api/v1/roasts/{id}/ — full detail, including sections/findings.
        DELETE /api/v1/roasts/{id}/ — hard-deletes the run (cascades to its
               sections/findings).
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - roasts
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoastRun'
          description: ''
    delete:
      operationId: roasts_destroy
      description: |-
        GET    /api/v1/roasts/{id}/ — full detail, including sections/findings.
        DELETE /api/v1/roasts/{id}/ — hard-deletes the run (cascades to its
               sections/findings).
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - roasts
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/submissions/{submission_id}/roasts/:
    get:
      operationId: submissions_roasts_list
      description: |-
        GET  /api/v1/submissions/{submission_id}/roasts/  — list this
             submission's roast runs (owner-scoped).
        POST /api/v1/submissions/{submission_id}/roasts/  — create a queued
             RoastRun and dispatch asynchronous processing. Never calls an
             AI provider directly — that only happens inside the Celery
             task (apps.roasts.tasks.process_roast_run).
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: path
        name: submission_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - roasts
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoastRunListList'
          description: ''
    post:
      operationId: submissions_roasts_create
      description: |-
        GET  /api/v1/submissions/{submission_id}/roasts/  — list this
             submission's roast runs (owner-scoped).
        POST /api/v1/submissions/{submission_id}/roasts/  — create a queued
             RoastRun and dispatch asynchronous processing. Never calls an
             AI provider directly — that only happens inside the Celery
             task (apps.roasts.tasks.process_roast_run).
      parameters:
      - in: path
        name: submission_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - roasts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoastRunCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RoastRunCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RoastRunCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoastRun'
          description: ''
...
components:
  schemas:
...
    ExtractionStatusEnum:
      enum:
      - queued
      - processing
      - completed
      - failed
      type: string
      description: |-
        * `queued` - Queued
        * `processing` - Processing
        * `completed` - Completed
        * `failed` - Failed
    IntensityEnum:
      enum:
      - gentle
      - sarcastic
      - brutal
      - nuclear
      type: string
      description: |-
        * `gentle` - Gentle
        * `sarcastic` - Sarcastic
        * `brutal` - Brutal
        * `nuclear` - Nuclear
    LanguageEnum:
      enum:
      - en
      - hi
      - hinglish
      type: string
      description: |-
        * `en` - English
        * `hi` - Hindi
        * `hinglish` - Hinglish
    PaginatedRoastRunListList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/RoastRunList'
        meta:
          type: object
          properties:
            count:
              type: integer
              example: 42
            next:
              type: string
              nullable: true
              format: uri
            previous:
              type: string
              nullable: true
              format: uri
            page_size:
              type: integer
              example: 20
...
    RoastFinding:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        category:
          type: string
          readOnly: true
        severity:
          allOf:
          - $ref: '#/components/schemas/SeverityEnum'
          readOnly: true
        title:
          type: string
          readOnly: true
        roast_text:
          type: string
          readOnly: true
        actual_feedback:
          type: string
          readOnly: true
        position:
          type: integer
          readOnly: true
        metadata:
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - actual_feedback
      - category
      - created_at
      - id
      - metadata
      - position
      - roast_text
      - severity
      - title
...
    RoastRun:
      type: object
      description: Full detail representation — includes nested sections/findings.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        submission:
          type: string
          format: uuid
          readOnly: true
        language:
          allOf:
          - $ref: '#/components/schemas/LanguageEnum'
          readOnly: true
        intensity:
          allOf:
          - $ref: '#/components/schemas/IntensityEnum'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/ExtractionStatusEnum'
          readOnly: true
        engine_version:
          type: string
          readOnly: true
        started_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        completed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        error_message:
          type: string
          readOnly: true
        summary:
          type: string
          readOnly: true
        final_verdict:
          type: string
          readOnly: true
        score:
          type: integer
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        sections:
          type: array
          items:
            $ref: '#/components/schemas/RoastSection'
          readOnly: true
        findings:
          type: array
          items:
            $ref: '#/components/schemas/RoastFinding'
          readOnly: true
      required:
      - completed_at
      - created_at
      - engine_version
      - error_message
      - final_verdict
      - findings
      - id
      - intensity
      - language
      - score
      - sections
      - started_at
      - status
      - submission
      - summary
      - updated_at
    RoastRunCreateRequest:
      type: object
      description: |-
        Only language/intensity come from the request body — the submission
        comes from the URL (/submissions/{submission_id}/roasts/) and is
        resolved (with ownership/status validation) by the view before this
        serializer is ever used. ChoiceField enforces "supported
        language"/"supported intensity" directly.
      properties:
        language:
          $ref: '#/components/schemas/LanguageEnum'
        intensity:
          $ref: '#/components/schemas/IntensityEnum'
      required:
      - intensity
      - language
    RoastRunList:
      type: object
      description: Lighter representation for listing a submission's roast runs.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        submission:
          type: string
          format: uuid
          readOnly: true
        language:
          allOf:
          - $ref: '#/components/schemas/LanguageEnum'
          readOnly: true
        intensity:
          allOf:
          - $ref: '#/components/schemas/IntensityEnum'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/ExtractionStatusEnum'
          readOnly: true
        engine_version:
          type: string
          readOnly: true
        started_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        completed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        error_message:
          type: string
          readOnly: true
        summary:
          type: string
          readOnly: true
        score:
          type: integer
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - completed_at
      - created_at
      - engine_version
      - error_message
      - id
      - intensity
      - language
      - score
      - started_at
      - status
      - submission
      - summary
...
    RoastSection:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        key:
          type: string
          readOnly: true
        title:
          type: string
          readOnly: true
        content:
          type: string
          readOnly: true
        position:
          type: integer
          readOnly: true
        metadata:
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - content
      - created_at
      - id
      - key
      - metadata
      - position
      - title
    SeverityEnum:
      enum:
      - info
      - low
      - medium
      - high
      - critical
      type: string
      description: |-
        * `info` - Info
        * `low` - Low
        * `medium` - Medium
        * `high` - High
        * `critical` - Critical
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT